chore(build): test auto deploy of PRs

+++ devdeploy.sh
@@ -1,9 +1,18 @@
 DEST="travis@fgpv.cloudapp.net:/home/travis/www"
 
-if [ -n $TRAVIS_TAG ]; then
-    DEST="$DEST/$TRAVIS_TAG/"
-else
-    DEST="$DEST/$TRAVIS_BRANCH/"
+if [ "$TRAVIS_REPO_SLUG" == "fgpv-vpgf/fgpv-vpgf" ]; then
+    openssl aes-256-cbc -k "$PW" -out ~/.ssh/id_rsa -in devkey.enc -d
+    echo -e "Host *\n\tStrictHostKeyChecking no\n" >> ~/.ssh/config
+    chmod 600 ~/.ssh/id_rsa
+    eval `ssh-agent -s`
+    ssh-add ~/.ssh/id_rsa
+
+    if [ -n $TRAVIS_TAG ]; then
+        DEST="$DEST/$TRAVIS_TAG/"
+    else
+        DEST="$DEST/$TRAVIS_BRANCH/"
+    fi
+
+    rsync -a --delete "build/" "$DEST"
 fi
 
-rsync -a --delete "build/" "$DEST"
